Cast enum to int.
Not sure if it's correct, but it's certainly not worse as before commit
48f8aa92514a5d0245b115263082b368f072c43e (Remove vm set state simcall)
which broke builds with clang.

+++ src/simix/smx_vm.c
@@ -96,7 +96,8 @@ static int __can_be_started(smx_host_t vm)
 void SIMIX_vm_start(smx_host_t ind_vm)
 {
   if (__can_be_started(ind_vm))
-	surf_resource_set_state(surf_workstation_resource_priv(ind_vm), SURF_VM_STATE_RUNNING);
+    surf_resource_set_state(surf_workstation_resource_priv(ind_vm),
+                            (int)SURF_VM_STATE_RUNNING);
   else
     THROWF(vm_error, 0, "The VM %s cannot be started", SIMIX_host_get_name(ind_vm));
 }
@@ -357,7 +358,8 @@ void SIMIX_vm_shutdown(smx_host_t ind_vm, smx_process_t issuer)
   }
 
   /* FIXME: we may have to do something at the surf layer, e.g., vcpu action */
-  surf_resource_set_state(surf_workstation_resource_priv(ind_vm), SURF_VM_STATE_CREATED);
+  surf_resource_set_state(surf_workstation_resource_priv(ind_vm),
+                          (int)SURF_VM_STATE_CREATED);
 }
 
 void SIMIX_pre_vm_shutdown(smx_simcall_t simcall, smx_host_t ind_vm)
